requirements.txt
requests
Pillow
python-telegram-bot
